+++ apps/dicompare/public/embed/DicompareReportRenderer.js
@@ -3,7 +3,7 @@
  * and generates standalone HTML for printing.
  *
  * Provided by dicompare-web for embedding in third-party tools.
- * See: https://github.com/astewartau/dicompare-web
+ * See: https://github.com/neurodesk/webapps/tree/main/apps/dicompare
  *
  * Usage:
  *   import { DicompareReportRenderer } from 'https://dicompare.neurodesk.org/embed/DicompareReportRenderer.js';